requirements.txt
anthropic>=0.18.0
boto3>=1.34.0
python-dotenv>=1.0.0
colorama>=0.4.6
flask>=3.0.0
flask-sock>=0.7.0
simple-websocket>=1.0.0
